refactor(flydsl): group GDR prefill kernels, standardize the K5 op test
The FlyDSL GDR prefill device kernels were scattered across kernels/, which
made it hard to tell which files belong to the prefill pipeline. Collect them
under kernels/gdr_prefill/ (matching the mqa_logits/mega_moe subpackage
layout), drop the now-redundant _opt suffix on the hidden-state kernel, and
leave re-export shims on the old module paths so existing imports keep working.
Public host entry points, the chunk_gdn_fwd_h_flydsl_opt kernel symbol and the
tuned CSV/env names are deliberately untouched, so tuned configs and AOT caches
still resolve.

Move the K5 prefill test next to its decode sibling in op_tests/ and rebuild it
on the aiter-op-test standard: @benchmark plus a run_perftest candidate loop
over flydsl/triton/hip, us + TFLOPS + TB/s + err per candidate, checkAllclose
against the fp32 torch reference, an arch allow-list in main() and a markdown
summary table. This replaces the pytest classes, the torch.profiler timing and
the hand-written speedup table. PrefillArgs and PREFILL_PARAMS stay module-level
because the BV tuner and the block benchmark import the catalog by path.
