+++ aiter/aot/flydsl/chunk_gdn_h.py
@@ -42,9 +42,7 @@ from aiter.aot.flydsl.common import (
     run_jobs_parallel,
 )
 from aiter.jit.core import AITER_CONFIGS
-from aiter.ops.flydsl.kernels.chunk_gated_delta_h_opt import (
-    compile_chunk_gated_delta_h_opt,
-)
+from aiter.ops.flydsl.kernels.gdr_prefill import compile_chunk_gated_delta_h
 from aiter.ops.flydsl.kernels.tensor_shim import _run_compiled
 
 CHUNK_GDN_H_AOT_ARCH_DEFAULT = "gfx950"
@@ -244,7 +242,7 @@ def _compile_to_cache(
         else int32_dummy
     )
 
-    launch_fn = compile_chunk_gated_delta_h_opt(
+    launch_fn = compile_chunk_gated_delta_h(
         K=K,
         V=V,
         BT=BT,
