+++ aiter/ops/flydsl/linear_attention_prefill_kernels.py
@@ -35,7 +35,7 @@ from ..triton._triton_kernels.gated_delta_rule.utils import (
     prepare_num_chunks,
     prepare_rebased_cu_seqlens,
 )
-from .kernels.gdn_prepare import compile_gdn_prepare
+from .kernels.gdr_prefill import compile_gdn_prepare
 from .kernels.tensor_shim import _run_compiled
 
 # log2(e); g pre-scaled by this constant lets the kernel use exp2(g) in
@@ -455,11 +455,9 @@ def _get_or_compile_opt(
             f"tiled-copy API), but got `{getattr(flydsl, '__version__', 'unknown')}`."
         )
 
-    from .kernels.chunk_gated_delta_h_opt import (
-        compile_chunk_gated_delta_h_opt,
-    )
+    from .kernels.gdr_prefill import compile_chunk_gated_delta_h
 
-    return compile_chunk_gated_delta_h_opt(
+    return compile_chunk_gated_delta_h(
         K=K,
         V=V,
         BT=BT,
